Почистила код? оформила классы в соответствии с рекомендациями
